Move unnecessary_cast to its own module
